src/requirements.txt
requests==2.31.0
pyTelegramBotAPI==4.13.0